+++ tests/End2EndTestCase.php
@@ -227,6 +227,7 @@ class End2EndTestCase extends IntegrationTestCase
 			'db.dsn' => 'pgsql:host=localhost;dbname=duoncms;user=duoncms;password=duoncms',
 			'path.root' => self::root(),
 			'path.public' => self::root() . '/public',
+			'path.views' => '/tests/Fixtures/templates',
 			'path.uploads' => self::root() . '/public/uploads',
 			'path.api' => '/api',
 			'path.panel' => '/panel',
@@ -271,19 +272,6 @@ class End2EndTestCase extends IntegrationTestCase
 		$plugin->node(\Duon\Cms\Tests\Fixtures\Node\TestWidget::class);
 		$plugin->node(\Duon\Cms\Tests\Fixtures\Node\TestDocument::class);
 		$plugin->node(\Duon\Cms\Tests\Fixtures\Node\TestMediaDocument::class);
-
-		$plugin->renderer('template', \Duon\Cms\Boiler\Renderer::class)->args(
-			dirs: self::root() . '/tests/Fixtures/templates',
-			autoescape: true,
-			trusted: [
-				Node::class,
-				Cms::class,
-				\Duon\Cms\Locales::class,
-				\Duon\Cms\Locale::class,
-				\Duon\Cms\Config::class,
-				Request::class,
-			],
-		);
 
 		return $plugin;
 	}